Refactor: firebase.ts 이동 및 관련 경로 수정
