+++ src/components/home/WikiPreview.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { Wiki } from "../wiki/WikiCommonType";
 import * as style from "./wikiPreviewtyle";
 import { doc, getDoc } from "firebase/firestore";
-import { db } from "../../../firebase";
+import { db } from "@/firebase/firebase";
 
 interface Props {
   wikiData: Wiki;
