+++ src/components/login/SignUp.tsx
@@ -1,6 +1,6 @@
 import React, { PropsWithChildren, useState } from "react";
 import { createUserWithEmailAndPassword } from "@firebase/auth";
-import { auth } from "../../../firebase";
+import { auth } from "@/firebase/firebase";
 
 import * as style from "./signupStyle";
