+++ src/firebase/services/wikiService.ts
@@ -1,5 +1,5 @@
 import { collection, getDocs, query, where } from "firebase/firestore";
-import { db } from "../../../firebase";
+import { db } from "@/firebase/firebase";
 
 export const hasChildWikis = async (wikiID: string) => {
   const wikisCollection = collection(db, "Wiki");
